requirements.txt
Flask==2.0.1
google-cloud-firestore==2.1.1
requests==2.18.0
pillow==8.2.0
numpy==1.20.3
